removes trafaret from servicelib[aiohttp]
